revert(netflix): 撤回共享 vm harness——它引入 3 条 S1523 且重复率收益未经测量
上一版把 node 测试的 DOM 桩 + vm 沙箱抽成 tools/browser-extension/scripts/
vm-content-harness.js，理由是审查**推测** Sonar 的 new_duplicated_lines_density
来自那批复制。我没有先测量就照做了，结果：

- Sonar 把 scripts/ 当生产代码，`vm.createContext` / `runInContext` 触发 3 条
  CRITICAL `javascript:S1523`（动态注入/执行代码），直接把 new_security_rating
  顶到 4（门限 1）。测试文件不吃这条规则，抽出来反而制造了新问题。
- 而那次 Sonar 的分析时间戳早于本 PR 的推送，5.2% 的重复到底来自哪里当时并没有
  证据；CPD 本来就只跑主源、不算测试文件，「测试里的复制导致重复率红」这个前提
  本身就没被验证过。

所以撤回：harness 放回 .test.js 里（那是它本来就该在的地方），行为修复一条不动。
真实的重复率等这一轮 Sonar 跑完拿数字再说，不再照推测改。

node 10/10、Dart 守卫 16/16、镜像字节一致守卫通过。
